wip - start create algo package and move files
