+++ libs/coin-algorand/src/api/algodv2.types.ts
