+++ libs/coin-algorand/src/bridge/js.ts
@@ -1,4 +1,6 @@
-import { makeAccountBridgeReceive } from "../../../bridge/jsHelpers";
+// FIXME: fix call to makeAccountBridgeReceive (dependency injection?)
+
+import { makeAccountBridgeReceive } from "@ledgerhq/coin-framework/bridge/jsHelpers";
 import type { AlgorandTransaction } from "../types";
 import { sync, scanAccounts } from "../js-synchronization";
 import {
@@ -7,17 +9,19 @@ import {
 } from "../js-prepareTransaction";
 import { estimateMaxSpendable } from "../js-estimateMaxSpendable";
 import { getTransactionStatus } from "../js-getTransactionStatus";
-import { signOperation } from "../js-signOperation";
+import { buildSignOperation } from "../js-signOperation";
 import { broadcast } from "../js-broadcast";
 import type { AccountBridge, CurrencyBridge } from "@ledgerhq/types-live";
 import { assignToAccountRaw, assignFromAccountRaw } from "../serialization";
 import { initAccount } from "../initAccount";
+import type { Transaction } from "../types";
 
 const receive = makeAccountBridgeReceive();
 
-const updateTransaction = (t, patch) => {
-  return { ...t, ...patch };
-};
+const updateTransaction = (t: Transaction, patch: Partial<Transaction>) => ({
+  ...t,
+  ...patch,
+});
 
 const preload = async () => Promise.resolve({});
