+++ libs/coin-algorand/src/errors.ts
@@ -1,0 +1,5 @@
+import { createCustomErrorClass } from "@ledgerhq/errors";
+
+export const AlgorandASANotOptInInRecipient = createCustomErrorClass(
+  "AlgorandASANotOptInInRecipient"
+);
