+++ libs/coin-algorand/src/js-broadcast.ts
@@ -1,5 +1,5 @@
 import type { Operation, SignedOperation } from "@ledgerhq/types-live";
-import { patchOperationWithHash } from "../../operation";
+import { patchOperationWithHash } from "@ledgerhq/coin-framework/operation";
 import { broadcastTransaction as apiBroadcast } from "./api";
 
 /**
