+++ libs/coin-algorand/src/js-estimateMaxSpendable.ts
@@ -1,6 +1,6 @@
 import { BigNumber } from "bignumber.js";
 import type { Account, AccountLike } from "@ledgerhq/types-live";
-import { getMainAccount } from "../../account";
+import { getMainAccount } from "@ledgerhq/coin-framework/account/index";
 import type { AlgorandTransaction, AlgorandAccount } from "./types";
 import { computeAlgoMaxSpendable } from "./logic";
 import { createTransaction } from "./js-prepareTransaction";
