+++ libs/coin-algorand/src/serialization.ts
