+++ libs/coin-algorand/src/specs.ts
