+++ libs/coin-algorand/src/speculos-deviceActions.ts
