+++ libs/ledger-live-common/package.json
@@ -127,6 +127,7 @@
     "@hashgraph/sdk": "^2.10.1",
     "@keplr-wallet/cosmos": "^0.9.16",
     "@keplr-wallet/proto-types": "^0.10.4",
+    "@ledgerhq/coin-algorand": "workspace:^",
     "@ledgerhq/coin-framework": "workspace:^",
     "@ledgerhq/coin-polkadot": "workspace:^",
     "@ledgerhq/compressjs": "github:LedgerHQ/compressjs#d9e8e4d994923e0ea76a32b97289bcccfe71b82e",
