+++ libs/ledgerjs/packages/errors/src/index.ts
@@ -26,6 +26,9 @@ export const CantOpenDevice = createCustomErrorClass("CantOpenDevice");
 export const CashAddrNotSupported = createCustomErrorClass(
   "CashAddrNotSupported"
 );
+export const ClaimRewardsFeesWarning = createCustomErrorClass(
+  "ClaimRewardsFeesWarning"
+);
 export const CurrencyNotSupported = createCustomErrorClass<
   { currencyName: string },
   LedgerErrorConstructor<{ currencyName: string }>
